# Precompute the Legendre normalization coefficients