--- a/numerics/legendre_body.hpp
+++ b/numerics/legendre_body.hpp
@@ -48,13 +48,6 @@ LegendrePolynomial() {
           std::make_index_sequence<degree_ + 1>>::coefficients);
 }
 
-inline double LegendreNormalizationFactor(int const degree, int const order) {
-  int const n = degree;
-  int const m = order;
-  return Sqrt((2 * n + 1) * (2 - (m == 0 ? 1 : 0)) /
-              static_cast<double>(FallingFactorial(n + m, 2 * m)));
-}
-
 }  // namespace internal_legendre
 }  // namespace numerics
 }  // namespace principia