--- a/numerics/max_abs_normalized_associated_legendre_functions_test.cc
+++ b/numerics/max_abs_normalized_associated_legendre_functions_test.cc
@@ -35,7 +35,7 @@ class MaxAbsNormalizedAssociatedLegendreFunctionTest : public testing::Test {
 TEST_F(MaxAbsNormalizedAssociatedLegendreFunctionTest, Polynomials) {
   for (int n = 0; n < MaxAbsNormalizedAssociatedLegendreFunction.rows; ++n) {
     EXPECT_THAT(MaxAbsNormalizedAssociatedLegendreFunction[n][0],
-                AlmostEquals(LegendreNormalizationFactor(n, 0), 0))
+                AlmostEquals(LegendreNormalizationFactor[n][0], 0))
         << "n = " << n;
   }
 }
@@ -43,7 +43,7 @@ TEST_F(MaxAbsNormalizedAssociatedLegendreFunctionTest, Polynomials) {
 TEST_F(MaxAbsNormalizedAssociatedLegendreFunctionTest, HighestOrder) {
   for (int n = 0; n < MaxAbsNormalizedAssociatedLegendreFunction.rows; ++n) {
     // TODO(phl): This should be
-    //  LegendreNormalizationFactor(n, n) * DoubleFactorial(2 * n - 1),
+    //  LegendreNormalizationFactor[n][n] * DoubleFactorial(2 * n - 1),
     // but that NaNs.
     EXPECT_THAT(MaxAbsNormalizedAssociatedLegendreFunction[n][n],
                 AlmostEquals(Sqrt((2 * n + 1) * (2 - (n == 0 ? 1 : 0)) /