… numerics/legendre_normalization_factor.mathematica.h
… physics/geopotential_test.cpp