Before:
```
Run on (4 X 3310 MHz CPU s)
CPU Caches:
  L1 Data 32K (x4)
  L1 Instruction 32K (x4)
  L2 Unified 262K (x4)
  L3 Unified 6291K (x1)
------------------------------------------------------------------------------
Benchmark                                       Time           CPU Iterations
------------------------------------------------------------------------------
BM_ComputeGeopotentialCpp/2                204554 ns     203362 ns       3452
BM_ComputeGeopotentialCpp/3                275033 ns     275333 ns       2493
BM_ComputeGeopotentialCpp/5                388112 ns     391089 ns       1795
BM_ComputeGeopotentialCpp/10              1131173 ns    1143845 ns        641
BM_ComputeGeopotentialF90/2                188696 ns     187752 ns       3739
BM_ComputeGeopotentialF90/3                233723 ns     234706 ns       2991
BM_ComputeGeopotentialF90/5                322053 ns     312836 ns       2244
BM_ComputeGeopotentialF90/10               791894 ns     792518 ns        748
BM_ComputeGeopotentialDistance/150000      198682 ns     200269 ns       3739
BM_ComputeGeopotentialDistance/500000       73931 ns      74750 ns       8974
BM_ComputeGeopotentialDistance/5000000      15876 ns      15680 ns      40792
```
After:
```
Run on (4 X 3310 MHz CPU s)
CPU Caches:
  L1 Data 32K (x4)
  L1 Instruction 32K (x4)
  L2 Unified 262K (x4)
  L3 Unified 6291K (x1)
------------------------------------------------------------------------------
Benchmark                                       Time           CPU Iterations
------------------------------------------------------------------------------
BM_ComputeGeopotentialCpp/2                191072 ns     191924 ns       3739
BM_ComputeGeopotentialCpp/3                262083 ns     265911 ns       2640
BM_ComputeGeopotentialCpp/5                360821 ns     359818 ns       1951
BM_ComputeGeopotentialCpp/10              1101162 ns    1095171 ns        641
BM_ComputeGeopotentialF90/2                189553 ns     191924 ns       3739
BM_ComputeGeopotentialF90/3                231744 ns     229490 ns       2991
BM_ComputeGeopotentialF90/5                312967 ns     312836 ns       2244
BM_ComputeGeopotentialF90/10               789215 ns     782614 ns        897
BM_ComputeGeopotentialDistance/150000      196732 ns     198843 ns       3452
BM_ComputeGeopotentialDistance/500000       71341 ns      72313 ns      11218
BM_ComputeGeopotentialDistance/5000000      15722 ns      15645 ns      44872
```